implements ldev authentication & role validation

+++ controllers/ldev_controller.py
@@ -1,16 +1,23 @@
 from flask import Blueprint, request, jsonify
+from flask_jwt_extended import jwt_required, get_jwt
 
 from services.ldev_service import create_ldev, list_ldevs, delete_ldev, get_ldev, update_ldev
 
 ldev_ = Blueprint('ldev', __name__, template_folder="./views", static_folder="./static", root_path="./")
 
 @ldev_.route('/', methods=['GET'])
+@jwt_required()
 def list_route():
     all_ldevs = list_ldevs()
     return jsonify([l.serialize() for l in all_ldevs]), 200
 
 @ldev_.route('/', methods=['POST'])
+@jwt_required()
 def create_route():
+    claims = get_jwt()
+    if claims.get('role') == 'user':
+        return {'error': 'Acesso negado'}, 403
+
     data = request.get_json() or {}
     required = ('name', 'latitude', 'longitude', 'locale_id')
     if not all(field in data for field in required):
@@ -25,12 +32,18 @@ def create_route():
     return jsonify(ldev.serialize()), 201
 
 @ldev_.route('/<string:ldev_id>', methods=['GET'])
+@jwt_required()
 def get_route(ldev_id):
     ldev = get_ldev(ldev_id)
     return jsonify(ldev.serialize()), 200
 
 @ldev_.route('/<string:ldev_id>', methods=['PATCH'])
+@jwt_required()
 def update_route(ldev_id):
+    claims = get_jwt()
+    if claims.get('role') == 'user':
+        return {'error': 'Acesso negado'}, 403
+
     ldev = get_ldev(ldev_id)
     data = request.get_json() or {}
     attrs = {}
@@ -45,7 +58,12 @@ def update_route(ldev_id):
     return jsonify(ldev.serialize()), 200
 
 @ldev_.route('/<string:ldev_id>', methods=['DELETE'])
+@jwt_required()
 def delete_route(ldev_id):
+    claims = get_jwt()
+    if claims.get('role') == 'user':
+        return {'error': 'Acesso negado'}, 403
+
     ldev = get_ldev(ldev_id)
     delete_ldev(ldev)
     return '', 204
